refactor: even nicer
